fix: Don't copy ublue-os key if image key exists

+++ modules/signing/signing.sh
@@ -25,7 +25,9 @@ if ! [ -f "$CONTAINER_DIR/policy.json" ]; then
     cp "$MODULE_DIRECTORY/signing/policy.json" "$CONTAINER_DIR/policy.json"
 fi
 
-cp "/usr/share/ublue-os/cosign.pub" "/usr/etc/pki/containers/$IMAGE_NAME.pub"
+if ! [ -f "/usr/etc/pki/containers/$IMAGE_NAME.pub" ]; then
+    cp "/usr/share/ublue-os/cosign.pub" "/usr/etc/pki/containers/$IMAGE_NAME.pub"
+fi
 
 POLICY_FILE="$CONTAINER_DIR/policy.json"
